requirements.txt
#requirements.txt
tweepy
praw
vaderSentiment
pandas
matplotlib
python-dotenv
streamlit
